Create linkmap.json and hwasm.py
Implement an ad hoc assembler+linker program 'hwasm' that takes 'hw.S' as input, and produces a flat-binary 'hw.bin', suitable to be loaded and executed using the x86 BIOS legacy boot. Upon requests, we're relaxing the requirement and allowing the ad hoc assembler to be written in Python. In this case, the executable is 'hwasm.py' and should be invoked like described, with one command-line argument. Please, do not use any magical library that already does the assembling for you.  Mind also that you'll need to update the rule that build 'hw.bin' and comment out the rule that would build 'hwasm'. The assembler may be a specific-purpose implementation that works only with the given input.